pyproject.toml
[project]
name = "jinnyai"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "py-cord[voice] (>=2.6.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "openai (>=1.63.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
